requirements.txt
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
Flask==2.0.1
importlib-metadata==4.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.1
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wincertstore==0.2
waitress==2.0.0
zipp==3.4.1
